requirements.txt
deta
flask
python-dotenv
pillow
asgiref
requests